[UPDATE] implementation, package.json, LICENSE, Makefile
