+++ lib/array.js
@@ -1,0 +1,33 @@
+'use strict';
+
+// MODULES //
+
+var ERFCINV = require( './number.js' );
+
+// INVERSE COMPLEMENTARY ERROR FUNCTION //
+
+/**
+* FUNCTION: erfcinv( out, arr )
+*	Computes the inverse complementary error function for each array element.
+*
+* @param {Number[]|Int8Array|Uint8Array|Uint8ClampedArray|Int16Array|Uint16Array|Int32Array|Uint32Array|Float32Array|Float64Array} out - output array
+* @param {Number[]|Int8Array|Uint8Array|Uint8ClampedArray|Int16Array|Uint16Array|Int32Array|Uint32Array|Float32Array|Float64Array} arr - input array
+* @returns {Number[]|Int8Array|Uint8Array|Uint8ClampedArray|Int16Array|Uint16Array|Int32Array|Uint32Array|Float32Array|Float64Array|Null} output array or null
+*/
+function erfcinv( y, x ) {
+	var len = x.length,
+		i;
+
+	if ( !len ) {
+		return null;
+	}
+	for ( i = 0; i < len; i++ ) {
+		y[ i ] = ERFCINV( x[ i ] );
+	}
+	return y;
+} // end FUNCTION erfcinv()
+
+
+// EXPORTS //
+
+module.exports = erfcinv;
